+++ src/azure-cli-core/azure/cli/core/commands/constants.py
@@ -7,7 +7,7 @@ from knack.parser import ARGPARSE_SUPPORTED_KWARGS
 
 
 CLI_COMMON_KWARGS = ['min_api', 'max_api', 'resource_type', 'operation_group',
-                     'custom_command_type', 'command_type']
+                     'custom_command_type', 'command_type', 'is_preview', 'preview_info']
 
 CLI_COMMAND_KWARGS = ['transform', 'table_transformer', 'confirmation', 'exception_handler',
                       'client_factory', 'operations_tmpl', 'no_wait_param', 'supports_no_wait', 'validator',