+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/commands.py
@@ -39,7 +39,7 @@ def load_command_table(self, _):
     # ACS base commands
     # TODO: When the first azure-cli release after January 31, 2020 is planned, add
     # `expiration=<CLI core version>` to the `self.deprecate()` args below.
-    deprecate_info = self.deprecate(redirect='aks')
+    deprecate_info = self.deprecate(redirect='aks', hide=True)
     with self.command_group('acs', container_services_sdk, deprecate_info=deprecate_info,
                             client_factory=cf_container_services) as g:
         g.custom_command('browse', 'acs_browse')
@@ -74,16 +74,16 @@ def load_command_table(self, _):
         g.custom_command('get-credentials', 'aks_get_credentials')
         g.command('get-upgrades', 'get_upgrade_profile', table_transformer=aks_upgrades_table_format)
         g.custom_command('install-cli', 'k8s_install_cli', client_factory=None)
-        g.custom_command('install-connector', 'k8s_install_connector')
+        g.custom_command('install-connector', 'k8s_install_connector', is_preview=True)
         g.custom_command('list', 'aks_list', table_transformer=aks_list_table_format)
-        g.custom_command('remove-connector', 'k8s_uninstall_connector')
+        g.custom_command('remove-connector', 'k8s_uninstall_connector', is_preview=True)
         g.custom_command('remove-dev-spaces', 'aks_remove_dev_spaces')
         g.custom_command('scale', 'aks_scale', supports_no_wait=True)
         g.custom_show_command('show', 'aks_show', table_transformer=aks_show_table_format)
         g.custom_command('upgrade', 'aks_upgrade', supports_no_wait=True,
                          confirmation='Kubernetes may be unavailable during cluster upgrades.\n' +
                          'Are you sure you want to perform this operation?')
-        g.custom_command('upgrade-connector', 'k8s_upgrade_connector')
+        g.custom_command('upgrade-connector', 'k8s_upgrade_connector', is_preview=True)
         g.custom_command('use-dev-spaces', 'aks_use_dev_spaces')
         g.wait_command('wait')