+++ src/command_modules/azure-cli-find/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.3.3
+++++++
+* Minor fixes
+
 0.3.2
 ++++++
 * Minor fixes