+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.4.4
++++++
+* Minor fixes
+
 0.4.3
 +++++
 * Minor fixes